+++ packages/askui-nodejs/src/core/model/test-case-dto/test-step.ts
@@ -1,7 +1,0 @@
-import { CustomElement } from './custom-element';
-
-export interface TestStep {
-  instruction: string
-  customElements?: CustomElement[]
-  secretText?: string | undefined
-}