+++ packages/askui-nodejs/src/core/reporting/instruction.ts
@@ -1,0 +1,8 @@
+import { CustomElement } from '@/core/model/custom-element';
+
+export interface Instruction {
+  readonly value: string;
+  readonly valueHumanReadable: string;
+  readonly customElements?: Readonly<Readonly<CustomElement>>[];
+  readonly secretText?: string | undefined;
+}