+++ packages/askui-nodejs/src/main.ts
@@ -1,3 +1,13 @@
 export { UiController } from './lib';
-export { UiControlClient, AnnotationLevel } from './execution';
+export { AnnotationLevel, UiControlClient } from './execution';
+export {
+  Instruction,
+  Reporter,
+  ReporterConfig,
+  Snapshot,
+  SnapshotDetailLevel,
+  Step,
+  StepStatus,
+  StepStatusEnd,
+} from './core/reporting';
 export { LogLevels } from './shared';